% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{asyncMeanCb}
\alias{asyncMeanCb}
\title{Compute mean on a background thread (callback result)}
\usage{
asyncMeanCb(data, callback)
}
\arguments{
\item{data}{numeric vector}

\item{callback}{an R function that will be called with the mean as its argument}
}
\description{
Like \code{asyncMean} but delivers the result to an R callback
  instead of printing it. Useful for testing and for integrating with promises.
}
\examples{
result <- NULL
asyncMeanCb(c(1, 2, 3), function(x) result <<- x)
later::run_now(2)
result
}
