% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laterexample-package.R
\docType{package}
\name{laterexample-package}
\alias{laterexample}
\alias{laterexample-package}
\title{laterexample: ADD TITLE}
\description{
Companion package to the 'later' vignette "Using later from C++". Demonstrates the later::BackgroundTask class using cpp4r.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://WEBSITE.COM}
  \item Report bugs at \url{https://github.com/USERNAME/PKGNAME/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
}

}
\keyword{internal}
